Move PATCH document_submissions to top level
